+++ app/src/main/java/com/compose/practical/ui/homeScreen/model/Destinations.kt
@@ -1,4 +1,4 @@
-package com.compose.practical.ui.homeScreen
+package com.compose.practical.ui.homeScreen.model
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
@@ -8,6 +8,7 @@ import androidx.compose.material.icons.filled.Person
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material.icons.filled.Star
 import androidx.compose.ui.graphics.vector.ImageVector
+import com.compose.practical.ui.homeScreen.model.Destinations.Add.path
 
 sealed class Destinations(
     val path: String,
@@ -30,6 +31,10 @@ sealed class Destinations(
                 Creation.path -> Creation
                 else -> Home
             }
+        }
+
+        val title = path.replaceFirstChar {
+            it.uppercase()
         }
 
     }
